layout stuff
I added a box to start the work of creating chat channels. I also started a page to show the chats. Next I will need to fill in the empty space with channels and chats that are user generated. After this I will make a basic login system using SQL